+++ tests/test_caching.py
@@ -11,6 +11,7 @@ import pytest
 from easypy.bunch import Bunch
 from easypy.caching import timecache, PersistentCache, cached_property, locking_cache
 from easypy.units import DAY
+from easypy.resilience import resilient
 
 _logger = getLogger(__name__)
 
@@ -202,9 +203,9 @@ def test_persistent_cache(persistent_cache_path):
     assert ps.get(TEST_KEY, default=None) is None, "Database was not cleared properly"
 
     # Expiration
-    ps = PersistentCache(persistent_cache_path, version=3, expiration=1 * DAY)
+    ps = PersistentCache(persistent_cache_path, version=3, expiration=.01)
     ps.set(TEST_KEY, TEST_VALUE)
-    ps.set("_PersistentCacheSignature", [3, time.time() - 2 * DAY])
+    time.sleep(0.011)
     assert ps.get(TEST_KEY, None) is None, "Database was not cleaned up on expiration"
 
 
@@ -259,3 +260,16 @@ def test_caching_gc_leaks(cache_decorator):
     del foo
     gc.collect()
     assert leaked() is None
+
+
+def test_resilient_between_timecaches():
+    class ExceptionLeakedThroughResilient(Exception):
+        pass
+
+    @timecache(1)
+    @resilient(acceptable=ExceptionLeakedThroughResilient, default='default')
+    @timecache(1)
+    def foo():
+        raise ExceptionLeakedThroughResilient()
+
+    assert foo() == 'default'